+++ apps/restconf/restconf_methods_patch.h
@@ -35,18 +35,14 @@
  */
 
 
-#ifndef _RESTCONF_METHODS_H_
-#define _RESTCONF_METHODS_H_
+#ifndef _RESTCONF_METHODS_PATCH_H_
+#define _RESTCONF_METHODS_PATCH_H_
 
 /*
  * Prototypes
  */
-int api_data_options(clicon_handle h, FCGX_Request *r);
-int api_data_put(clicon_handle h, FCGX_Request *r, char *api_path, 
-		 cvec *pcvec, int pi, 
-		 cvec *qvec, char *data,
-		 int pretty, int use_xml, int parse_xml);
-int api_data_delete(clicon_handle h, FCGX_Request *r, char *api_path, int pi,
-		    int pretty, int use_xml);
+int api_data_patch(clicon_handle h, FCGX_Request *r, char *api_path, 
+		   cvec *pcvec, int pi, 
+		   cvec *qvec, char *data);
 
-#endif /* _RESTCONF_METHODS_H_ */
+#endif /* _RESTCONF_METHODS_PATCH_H_ */