+++ lib/clixon/clixon_json.h
@@ -41,6 +41,7 @@
 /*
  * Prototypes
  */
+int json2xml_decode(cxobj *x, cxobj **xerr);
 int xml2json_cbuf(cbuf *cb, cxobj *x, int pretty);
 int xml2json_cbuf_vec(cbuf *cb, cxobj **vec, size_t veclen, int pretty);
 int xml2json(FILE *f, cxobj *x, int pretty);