+++ test/test_xml.sh
@@ -8,6 +8,8 @@ s="$_" ; . ./lib.sh || if [ "$s" = $0 ]; then exit 0; else return 0; fi
 
 : ${clixon_util_xml:=clixon_util_xml -o} # -o is output
 
+sleep 1 # mysterious fail, maybe this helps?
+
 new "xml parse"
 expecteof "$clixon_util_xml" 0 "<a><b/></a>" "^<a><b/></a>$"